LayoutTests/platform/chromium-mac/plugins/plugin-clip-subframe-expected.txt

CONSOLE MESSAGE: line 5: NPP_SetWindow: NULL window, Rect {0, 0, 300, 150}, Clip Rect {65518, 65318, 65518, 65318}, Type 2
CONSOLE MESSAGE: line 5: NPP_SetWindow: NULL window, Rect {0, 0, 300, 150}, Clip Rect {0, 0, 90, 150}, Type 2
